캐릭터 Hp, Mp 표시를 위한 위젯 추가
Attribute 를 int 에서 float 로 변경
GameMode 를 통해 Status 표시를 위한 위젯이 준비 완료 된 후 요소를 전달하여 신뢰성 보장하도록 코드 작성

+++ Source/LshPF/Public/Data/LshPF_DataTableRow.h
@@ -10,23 +10,23 @@ struct FEnemyAttribute : public FTableRowBase
 	GENERATED_USTRUCT_BODY()
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute")
-	int32 BaseMaxHealth = 0;
+	float BaseMaxHealth = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute")
-	int32 BaseMaxMana = 0;
+	float BaseMaxMana = 0;
 	
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute")
-	int32 BaseSpeed = 0;
+	float BaseSpeed = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute")
-	int32 BaseAttack = 0;
+	float BaseAttack = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute")
-	int32 BaseDefence = 0;
+	float BaseDefence = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute")
-	int32 BaseAbilityAttack = 0;
+	float BaseAbilityAttack = 0;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Attribute")
-	int32 BaseAbilityDefence = 0;
+	float BaseAbilityDefence = 0;
 };
